% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxt4.R
\docType{package}
\name{Maxt4}
\alias{Maxt4}
\title{Find test statistic value and corresponding critical value}
\description{
This package finds test statistic value and corresponding
critical value of Max-T test for testing
equality of means against ordered alternatives in one-way way ANOVA
model with heteroscedasticity when number of groups is 4.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
