% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxt4crit.R
\name{Maxt4crit}
\alias{Maxt4crit}
\title{Find test statistic value and corresponding critical value}
\usage{
Maxt4crit(g1, g2, g3, g4, alpha)
}
\arguments{
\item{g1}{a real vector}

\item{g2}{a real vector}

\item{g3}{a real vector}

\item{g4}{a real vector}

\item{alpha}{a real number between 0 and 1}
}
\value{
a numerical vector
}
\description{
More detailed description
}
\examples{
g1=c(1,2,4,1,1,3,4,0,2,4,1,3,5,2,1,4)
g2=c(0.4,2,-4,0.9,4.6,8,9.8,9,4,5,6,3)
g3=c(0,1,2,5,7,3,7.9,3,2)
g4=c(0,-3,-4,5,1,3,5,7,2,1,2,0)
Maxt4crit(g1,g2,g3,g4,0.05)

}
